feat: Use HashRouter for development and BrowserRouter for production
This change dynamically selects the router based on the environment.
It uses HashRouter when not on the production domain ('aprende.marketing')
to facilitate local development and deployment on platforms that might
not support client-side routing out-of-the-box. BrowserRouter is used
on the production domain.

Additionally, the Dockerfile has been updated to support a more robust
build and run process for the application.